cloud_security_scout/pyproject.toml
[project]
name = "cloud-security-scout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.43",
    "detect-secrets>=1.5.0",
    "pre-commit>=4.2.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "tabulate>=0.9.0",
]
